pyproject.toml
[tool.poetry]
name = "xjet"
version = "0.1.1"
description = "Python SDK for t.me/xJetSwapBot"
readme = "README.md"
authors = [
    "delpydoc <delpydoc@proton.me>",
]
maintainers = []
repository = "https://github.com/xJetLabs/python-sdk"
homepage = "https://github.com/xJetLabs/python-sdk"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
]
keywords = ["ton", "the open network", "xjet", "xjetswap", "sdk"]

[tool.setuptools.package-data]
xjet = []

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
httpx = ">=0.23.1"
PyNaCl = ">=1.5.0"

[build-system]
requires = ["poetry>=1.2.0"]
build-backend = "poetry.core.masonry.api"
